refactor(Controller, Dealer): replace extension functions with internal service methods - Remove Dealer.dealCards() and Players.dealCards() extension functions - Add serviceDealer(), servicePlayers(), and dealCards() to Dealer - Refactor tests according to above changes
