+++ src/test/kotlin/blackjack/DealerTest.kt
@@ -1,7 +1,11 @@
 package blackjack
 
 import blackjack.model.Card
+import blackjack.model.CardDeck
 import blackjack.model.Dealer
+import blackjack.model.HandCards
+import blackjack.model.Player
+import blackjack.model.Players
 import blackjack.model.Rank
 import blackjack.model.Suit
 import org.assertj.core.api.Assertions.assertThat
@@ -9,9 +13,15 @@ import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 
 class DealerTest {
+
+    private val deck = CardDeck()
+    private val playerCards = mutableListOf(Card(Rank.ACE, Suit.DIAMONDS), Card(Rank.KING, Suit.DIAMONDS))
+    private val player = Player("Lisa", handCards = HandCards(playerCards))
+
     @Test
     fun `dealer should draw when 16`() {
-        val dealer = Dealer()
+
+        val dealer = Dealer(deck = deck, players = Players(listOf(player)))
         dealer.handCards.add(Card(Rank.ACE, Suit.DIAMONDS))
         dealer.handCards.add(Card(Rank.FIVE, Suit.DIAMONDS))
         assertTrue { dealer.shouldDraw() }
@@ -19,7 +29,7 @@ class DealerTest {
 
     @Test
     fun `dealer should not draw when 17`() {
-        val dealer = Dealer()
+        val dealer = Dealer(deck = deck, players = Players(listOf(player)))
         dealer.handCards.add(Card(Rank.ACE, Suit.DIAMONDS))
         dealer.handCards.add(Card(Rank.SIX, Suit.DIAMONDS))
         assertThat(dealer.shouldDraw()).isFalse()
@@ -27,7 +37,7 @@ class DealerTest {
 
     @Test
     fun `dealer hand cards reduce total when necessary`() {
-        val dealer = Dealer()
+        val dealer = Dealer(deck = deck, players = Players(listOf(player)))
         dealer.handCards.add(Card(Rank.KING, Suit.DIAMONDS))
         dealer.handCards.add(Card(Rank.ACE, Suit.DIAMONDS))
         dealer.handCards.add(Card(Rank.FIVE, Suit.DIAMONDS))
