+++ src/test/kotlin/blackjack/ResultsTest.kt
@@ -1,6 +1,7 @@
 package blackjack
 
 import blackjack.model.Card
+import blackjack.model.CardDeck
 import blackjack.model.Dealer
 import blackjack.model.HandCards
 import blackjack.model.Player
@@ -14,13 +15,15 @@ import org.junit.jupiter.api.Test
 
 class ResultsTest {
 
+    private val deck = CardDeck()
+
     @Test
     fun `test dealer is not busted, player is not busted, DEALER WINS`() {
         val playersCards = mutableListOf(Card(Rank.ACE, Suit.DIAMONDS), Card(Rank.NINE, Suit.DIAMONDS)) // total 20
         val dealersCards = mutableListOf(Card(Rank.ACE, Suit.DIAMONDS), Card(Rank.KING, Suit.DIAMONDS)) // total 21
-        val dealer = Dealer(handCards = HandCards(dealersCards))
         val player = Player("test", HandCards(playersCards))
         val players = Players(listOf(player))
+        val dealer = Dealer(deck = deck, players = players, handCards = HandCards(dealersCards))
         val evaluator = ResultEvaluator(players, dealer)
         val results = evaluator.calculateResults()
         OutputView.displayAllCardsMessage(dealer)
@@ -36,9 +39,9 @@ class ResultsTest {
     fun `test dealer is not busted, player is not busted, PLAYER WINS`() {
         val playersCards = mutableListOf(Card(Rank.ACE, Suit.DIAMONDS), Card(Rank.KING, Suit.DIAMONDS)) // total 21
         val dealersCards = mutableListOf(Card(Rank.ACE, Suit.DIAMONDS), Card(Rank.NINE, Suit.DIAMONDS)) // total 20
-        val dealer = Dealer(handCards = HandCards(dealersCards))
         val player = Player("test", HandCards(playersCards))
         val players = Players(listOf(player))
+        val dealer = Dealer(deck = deck, players = players, handCards = HandCards(dealersCards))
         val evaluator = ResultEvaluator(players, dealer)
         val results = evaluator.calculateResults()
         OutputView.displayAllCardsMessage(dealer)
@@ -54,9 +57,9 @@ class ResultsTest {
     fun `test dealer is not busted, player is not busted, DRAW`() {
         val playersCards = mutableListOf(Card(Rank.TEN, Suit.DIAMONDS), Card(Rank.NINE, Suit.DIAMONDS)) // total 19
         val dealersCards = mutableListOf(Card(Rank.NINE, Suit.CLUBS), Card(Rank.TEN, Suit.SPADES)) // total 19
-        val dealer = Dealer(handCards = HandCards(dealersCards))
         val player = Player("test", HandCards(playersCards))
         val players = Players(listOf(player))
+        val dealer = Dealer(deck = deck, players = players, handCards = HandCards(dealersCards))
         val evaluator = ResultEvaluator(players, dealer)
         val results = evaluator.calculateResults()
         OutputView.displayAllCardsMessage(dealer)
@@ -72,9 +75,9 @@ class ResultsTest {
     fun `test dealer is not busted, player is busted, DEALER WINS`() {
         val playersCards = mutableListOf(Card(Rank.KING, Suit.HEARTS), Card(Rank.NINE, Suit.CLUBS), Card(Rank.FIVE, Suit.SPADES)) // total 24 busted
         val dealersCards = mutableListOf(Card(Rank.TEN, Suit.DIAMONDS), Card(Rank.NINE, Suit.DIAMONDS)) // total 19
-        val dealer = Dealer(handCards = HandCards(dealersCards))
         val player = Player("test", HandCards(playersCards))
         val players = Players(listOf(player))
+        val dealer = Dealer(deck = deck, players = players, handCards = HandCards(dealersCards))
         val evaluator = ResultEvaluator(players, dealer)
         val results = evaluator.calculateResults()
         OutputView.displayAllCardsMessage(dealer)
@@ -90,9 +93,9 @@ class ResultsTest {
     fun `test dealer is busted, player is not busted, PLAYER WINS`() {
         val playersCards = mutableListOf(Card(Rank.NINE, Suit.CLUBS), Card(Rank.KING, Suit.SPADES)) // total 19
         val dealersCards = mutableListOf(Card(Rank.KING, Suit.HEARTS), Card(Rank.NINE, Suit.DIAMONDS), Card(Rank.FIVE, Suit.CLUBS)) // total 24 busted
-        val dealer = Dealer(handCards = HandCards(dealersCards))
         val player = Player("test", HandCards(playersCards))
         val players = Players(listOf(player))
+        val dealer = Dealer(deck = deck, players = players, handCards = HandCards(dealersCards))
         val evaluator = ResultEvaluator(players, dealer)
         val results = evaluator.calculateResults()
         OutputView.displayAllCardsMessage(dealer)
@@ -108,9 +111,9 @@ class ResultsTest {
     fun `test dealer is busted, player is busted, LOSS FOR PLAYER, DEALER WINS`() {
         val playersCards = mutableListOf(Card(Rank.KING, Suit.HEARTS), Card(Rank.NINE, Suit.CLUBS), Card(Rank.FIVE, Suit.SPADES)) // 24 busted
         val dealersCards = mutableListOf(Card(Rank.KING, Suit.DIAMONDS), Card(Rank.NINE, Suit.HEARTS), Card(Rank.SIX, Suit.CLUBS)) // 25 busted
-        val dealer = Dealer(handCards = HandCards(dealersCards))
         val player = Player("test", HandCards(playersCards))
         val players = Players(listOf(player))
+        val dealer = Dealer(deck = deck, players = players, handCards = HandCards(dealersCards))
         val evaluator = ResultEvaluator(players, dealer)
         val results = evaluator.calculateResults()
         OutputView.displayAllCardsMessage(dealer)
